LPD-41245 Also move and rename EhcacheValue to expiry/EhcacheExpiryValue, as this class is completely added to support per-entry TTL
